refactor: extract 6 concerns from App.vue into helpers
Moves ~100 lines of self-contained logic out of App.vue (1283→1252)
into focused modules:

src/utils/session/sessionHelpers.ts:
  pushResult, pushErrorMessage, beginUserTurn,
  appendToLastAssistantText

src/utils/session/seedRoleDefault.ts:
  maybeSeedRoleDefault (source-manager preload)

src/utils/tools/dedup.ts:
  deduplicateResults (sidebar consecutive-tool collapse)

src/composables/useKeyNavigation.ts:
  handleCanvasKeydown + handleKeyNavigation

All call sites in App.vue now import instead of defining inline.
No behavior change — the functions have identical logic, just moved
to testable, reusable locations.
